Refactored moran analysis
