+++ lib/node/index.js
@@ -2,7 +2,7 @@
 
 var nodes = {
     Buffer: require('./nodes/buffer'),
-    Moran: require('./nodes/moran'),
+    Moran: require('./nodes/moran/moran'),
     PointInPolygon: require('./nodes/point-in-polygon'),
     PopulationInArea: require('./nodes/population-in-area'),
     Source: require('./nodes/source'),
